Search for BUILD.bazel before BUILD
The current search order fails on case-insensitive filesystems if
a normal file called "build" exist in the path, because it'll only
fall back to BUILD.bazel if "build" is a directory. This will also
cause issues if the build dir isn't created yet. If there's a
BUILD.bazel file it's safe to assume the user has decided to use
that naming for a reason, so we should check for that first and
prefer it.